requirements.txt
django>=1.10.0
django-oauth2-provider>=0.2.6.1
django-oauth-toolkit>=0.10.0
django-cors-headers>=1.1.0
ldap3>=1.4.0
tqdm>=4.8.4
flake8>=3.0.4
djangorestframework>=3.5.3

# For student image import / manipulation
requests>=2.11.1
pillow>=3.4.2

# For more filters in the REST API
django-filter>=0.15.3

# For the extended filter language
pre-js-py==1.1.0

# For testing pep
pep8 >= 1.7.0

# For using Shibboleth / SAML as an upstream ID provider
python3-saml == 1.2.1
social-auth-app-django == 1.2.0

# mysqlclient>=1.3.9